+++ src/Shrun/Notify/DBus.hs
@@ -57,8 +57,8 @@ shrunToDBus :: ShrunNote -> Note
 shrunToDBus shrunNote =
   DBusN.Note
     { appName = "Shrun",
-      summary = unpack $ shrunNote ^. #summary % #unUnlinedText,
-      body = Just . DBusN.Text . T.unpack $ shrunNote ^. #body % #unUnlinedText,
+      summary = unpack $ shrunNote ^. #summary % #unNotifyMessage,
+      body = Just . DBusN.Text . T.unpack $ shrunNote ^. #body % #unNotifyMessage,
       appImage = Nothing,
       hints = [Urgency (shrunNote ^. #urgency)],
       expiry,